requirements.txt
djangorestframework
pillow
aiogram==2.25.1
requests


